Use 'i16' for 'MatrixPosition'
I would remove 'MatrixPosition' entirely, as I've now realized that the translation component of 'Transform' is meant for this usage, except that it uses 'f32's, which are susceptible to floating point imprecision. Keeping that in mind, I would use 'i8'/'u8', but that may not be sufficient (as I plan to allow the user to choose the playing board dimensions). I may use 'u16' instead and only cast to and from 'i16' while bounds haven't been applied.

+++ src/matrix.rs
@@ -1,11 +1,11 @@
 pub struct Matrix {
-    pub width: i32,
-    pub height: i32,
+    pub width: i16,
+    pub height: i16,
 }
 
 // Holds a block's position within a tetromino for rotation
 #[derive(Debug, Clone, Copy)]
 pub struct MatrixPosition {
-    pub x: i32,
-    pub y: i32,
+    pub x: i16,
+    pub y: i16,
 }
